Isolate search module types
