+++ src/models/search-module/raw-search-module.ts
@@ -1,0 +1,9 @@
+export interface RawSearchModule {
+	Name: string;
+	Info: string;
+	Examples: Array<string>;
+
+	FrontendOnly: boolean;
+	Collapsing: boolean;
+	Sorting: boolean;
+}
